change avg sell time to int in api

+++ Models/PriceEstimate.cs
@@ -10,7 +10,7 @@ namespace Coflnet.Sky.Sniper.Models
         public string MedianKey { get; set; }
         public string ItemKey { get; set; }
         public byte Volatility { get; set; }
-        public short AvgSellTime { get; set; }
+        public int AvgSellTime { get; set; }
         public ReferencePrice LastSale { get; set; }
     }
 }
